Move progressive-enhancement, transition-context into 'ui' folder.
