+++ frontend/lib/ui/confetti-loadable.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { ConfettiProps } from './confetti';
-import { SimpleProgressiveEnhancement } from '../progressive-enhancement';
+import { SimpleProgressiveEnhancement } from './progressive-enhancement';
 import loadable from '@loadable/component';
 
 const LoadableConfetti = loadable(() => import('./confetti'), {
